# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.CreateAttractionRequest:
    properties:
      category:
        type: string
      contact_info:
        maxLength: 200
        type: string
      description:
        maxLength: 1000
        type: string
      estimated_visit_duration_minutes:
        type: integer
      hours_of_operation:
        maxLength: 200
        type: string
      location:
        maxLength: 300
        minLength: 1
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - category
    - location
    - name
    type: object
...
  controllers.CreateTransitionRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      hours_of_operation:
        maxLength: 200
        type: string
      mode:
        type: string
      route_notes:
        maxLength: 500
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - mode
    - title
    type: object
...
  middleware.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  middleware.SwaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorInfo'
      success:
        example: false
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@travellink.local
    name: API Support
    url: http://www.swagger.io/support
  description: Travel Linked List API - Browse, create, and share travel plans
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TravelLink API
  version: "1.0"
paths:
...
  /nodes:
    get:
      description: Get paginated list of approved nodes (attractions and/or transitions).
        Can filter by type. (public endpoint)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: limit
        type: integer
      - description: Node type filter
        enum:
        - attraction
        - transition
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Nodes list with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      summary: List available nodes
      tags:
      - nodes
  /nodes/{id}:
    delete:
      description: Admin can delete a node (admin only)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Delete a node
      tags:
      - admin
    get:
      description: Retrieve detailed information about a node (attraction or transition).
        Details include full information for the specific node type. (public endpoint)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      summary: Get node details
      tags:
      - nodes
...
  /nodes/attraction:
    post:
      consumes:
      - application/json
      description: Create a new user-generated attraction node. Requires admin approval
        before public use. Traveller or admin only.
      parameters:
      - description: Attraction creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateAttractionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attraction node created (pending approval)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "403":
          description: Only traveller or admin can create nodes
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Create an attraction node
      tags:
      - nodes
  /nodes/transition:
    post:
      consumes:
      - application/json
      description: |-
        Create a new user-generated transition node (journey between attractions). Requires admin approval before public use. Traveller or admin only.
        Transition nodes represent movement between attractions with modes like walking, car, bus, train, bike, taxi, or flight.
        Fields: title (service/line identifier like "Bus Line 5"), mode (transportation type), hours (operating hours if applicable), description, route_notes, distance
      parameters:
      - description: Transition creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateTransitionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transition node created (pending approval)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "403":
          description: Only traveller or admin can create nodes
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Create a transition node
      tags:
      - nodes
...
securityDefinitions:
  Bearer:
    description: 'Enter JWT token in the format: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"